+++ site/components.d.ts
@@ -10,8 +10,11 @@ declare module 'vue' {
     CreateRepo: typeof import('./src/components/CreateRepo.vue')['default']
     CreateStorage: typeof import('./src/components/CreateStorage.vue')['default']
     CreateUser: typeof import('./src/components/CreateUser.vue')['default']
+    DeployRepo: typeof import('./src/components/repo/update/DeployRepo.vue')['default']
     EditItem: typeof import('./src/components/repo/edit/EditItem.vue')['default']
     EditMenu: typeof import('./src/components/repo/edit/EditMenu.vue')['default']
+    FrontendRepo: typeof import('./src/components/repo/update/FrontendRepo.vue')['default']
+    GeneralRepo: typeof import('./src/components/repo/update/GeneralRepo.vue')['default']
     Item: typeof import('./src/components/admin/Item.vue')['default']
     Login: typeof import('./src/components/nav/Login.vue')['default']
     MavenProjectCopy: typeof import('./src/components/project/types/maven/copy/MavenProjectCopy.vue')['default']
@@ -24,10 +27,11 @@ declare module 'vue' {
     ProjectBadge: typeof import('./src/components/project/badge/ProjectBadge.vue')['default']
     Repositories: typeof import('./src/components/Repositories.vue')['default']
     RepositoryBadge: typeof import('./src/components/repo/badge/RepositoryBadge.vue')['default']
+    SecurityRepo: typeof import('./src/components/repo/update/SecurityRepo.vue')['default']
     Settings: typeof import('./src/components/Settings.vue')['default']
     SideBar: typeof import('./src/components/SideBar.vue')['default']
     Storages: typeof import('./src/components/Storages.vue')['default']
-    UpdateRepo: typeof import('./src/components/repo/UpdateRepo.vue')['default']
+    UpdateRepo: typeof import('./src/components/repo/update/UpdateRepo.vue')['default']
     UpdateStorage: typeof import('./src/components/UpdateStorage.vue')['default']
     UpdateUser: typeof import('./src/components/UpdateUser.vue')['default']
     UserEditMenu: typeof import('./src/components/user/edit/UserEditMenu.vue')['default']
