+++ site/src/backend/api/admin/Repository.ts
@@ -1,8 +1,9 @@
-import {BasicResponse} from "../../Response";
+import { BasicResponse, Frontend } from "../../Response";
 import http from "@/http-common";
-import {Err, Ok} from "ts-results";
-import {createAPIError, INTERNAL_ERROR, NOT_AUTHORIZED,} from "../../NitroRepoAPI";
-import {Repository} from "@/backend/Response";
+import { Err, Ok } from "ts-results";
+import { createAPIError, INTERNAL_ERROR, NOT_AUTHORIZED, } from "../../NitroRepoAPI";
+import { Repository } from "@/backend/Response";
+import { stringifyQuery } from "vue-router";
 
 export async function createNewRepository(
   name: string,
@@ -186,10 +187,11 @@ export async function updateFrontend(
   pageProvider: string,
   token: string
 ) {
+// Manually converting data to JSON because JSON.stringify is convering booleans to strings?
   return http
     .patch(
       "/api/admin/repository/" + id + "/modify/settings/frontend",
-      { enabled: enabled, page_provider: pageProvider },
+      `{"page_provider":"${pageProvider}","enabled":${enabled}}`,
       {
         headers: {
           Authorization: "Bearer " + token,
@@ -232,9 +234,9 @@ export async function setVisibility(
   return http
     .patch(
       "/api/admin/repository/" +
-        id +
-        "/modify/security/visibility/" +
-        visibility,
+      id +
+      "/modify/security/visibility/" +
+      visibility,
       {},
       {
         headers: {
@@ -317,13 +319,13 @@ export async function addOrRemoveReadersOrDeployers(
   return http
     .patch(
       "/api/admin/repository/" +
-        id +
-        "/modify/security/" +
-        what +
-        "/" +
-        action +
-        "/" +
-        user,
+      id +
+      "/modify/security/" +
+      what +
+      "/" +
+      action +
+      "/" +
+      user,
       {},
       {
         headers: {
