+++ site/src/main.ts
@@ -6,6 +6,7 @@ import {VueCookieNext} from "vue-cookie-next";
 import Notifications from "@kyvg/vue3-notification";
 import {createMetaManager} from "vue-meta";
 import "@/styles/app.css";
+import "@/styles/form.css";
 import "boxicons/css/boxicons.min.css";
 import "boxicons/dist/boxicons.js";
 import VueUploadComponent from "vue-upload-component";
