+++ site/src/styles/form.css
@@ -1,0 +1,81 @@
+.settingContent {
+  @apply max-w-lg;
+  @apply mx-auto;
+}
+
+.settingBox {
+  @apply md:w-1/2;
+  @apply px-3;
+}
+
+.nitroLabel {
+  @apply block;
+  @apply uppercase;
+  @apply tracking-wide;
+  @apply text-white;
+  @apply text-xs;
+  @apply font-bold;
+  @apply text-left;
+  @apply my-3;
+}
+.nitroSelectBox {
+  @apply appearance-none;
+  @apply block;
+  @apply w-full;
+  @apply bg-gray-200;
+  @apply text-gray-700;
+  @apply border;
+  @apply border-gray-200;
+  @apply rounded;
+  @apply py-3;
+  @apply px-4;
+  @apply leading-tight;
+  @apply focus:outline-none;
+  @apply focus:bg-white;
+  @apply focus:border-gray-500;
+}
+.nitroTextInput {
+  @apply appearance-none;
+  @apply block;
+  @apply w-full;
+  @apply bg-gray-200;
+  @apply text-gray-700;
+  @apply border;
+  @apply border-gray-200;
+  @apply rounded;
+  @apply py-3;
+  @apply px-4;
+  @apply leading-tight;
+  @apply focus:outline-none;
+  @apply focus:bg-white;
+  @apply focus:border-gray-500;
+}
+
+.disabled {
+  @apply bg-gray-300 !important;
+  @apply text-gray-700 !important;
+  @apply border-gray-800 !important;
+
+}
+.settingHeader{
+  @apply text-slate-100;
+  @apply shadow-sm;
+  @apply shadow-slate-400;
+  @apply  m-4;
+  @apply py-2;
+  @apply  text-left;
+  @apply font-bold;
+  
+}
+.settingSection{
+  @apply flex;
+  @apply flex-wrap;
+  @apply  mb-6
+}
+.nitroButton{
+  @apply bg-slate-800;
+  @apply p-4 my-3;
+  @apply rounded-md;
+  @apply cursor-pointer;
+  @apply text-white;
+}
